+++ platform/view/services/comm/p2p_test.go
@@ -8,7 +8,7 @@ package comm
 
 import (
 	"context"
-	"fmt"
+	assert2 "github.com/hyperledger-labs/fabric-smart-client/platform/view/services/assert"
 	"io/ioutil"
 	"sync"
 	"testing"
@@ -115,11 +115,11 @@ func setupTwoNodesFromFiles(t *testing.T) (*P2PNode, *P2PNode, string, string) {
 	nodeEndpoint := "/ip4/127.0.0.1/tcp/1235"
 
 	var bootstrapNode, node *P2PNode
-	assert.NoError(t, Retry(3, 1*time.Second, func() error {
+	assert2.EventuallyWithRetry(t, 3, 1*time.Second, func() error {
 		var err error
 		bootstrapNode, node, err = setupTwoNodes(t, bootstrapNodeID, bootstrapNodeEndpoint, nodeID, nodeEndpoint, &PrivateKeyFromFile{bootstrapNodeSK}, &PrivateKeyFromFile{nodeSK})
 		return err
-	}), "failed to setup two nodes")
+	}, "failed to setup two nodes")
 
 	return bootstrapNode, node, bootstrapNodeID, nodeID
 }
@@ -273,19 +273,4 @@ func SessionsForMPCTestRound(t *testing.T, bootstrapNode, node *P2PNode, bootstr
 
 	bootstrapNode.Stop()
 	node.Stop()
-}
-
-func Retry(attempts int, sleep time.Duration, f func() error) (err error) {
-	for i := 0; i < attempts; i++ {
-		if i > 0 {
-			time.Sleep(sleep)
-			sleep *= 2
-		}
-
-		err = f()
-		if err == nil {
-			return nil
-		}
-	}
-	return fmt.Errorf("no luck after %d attempts: last error: %v", attempts, err)
 }
